FW/Mac: fix build: Add support for Mach-O sections
First, the __attribute__((section)) must include the segment name, which
for read-write content is "__DATA".

Second, the linker doesn't automatically generate the start-stop
symbols, so we add a .S file to create them for us. We must ensure this
file is one of the first to be linked, so we move everything into
link_whole.