FW/CPUID: add support for Linux 5.16 state saving requests
With this implementation, we will only issue the new system call if
CPUID indicates AMX support, in which case we ignore XCR0. This is
designed for the likely scenario of AMX-capable kernels being used for
AMX-capable hardware. If, instead, we're running on a system that fails
to enable all CPU features, they won't begrudge us a few more CPU cycles
to find out.

We should also support the case in which the 5.16+ Linux kernel does
support AMX but refuses to offer it to us. For that case, we expect the
ARCH_GET_XCOMP_SUPP call to tell us "no AMX".