FW: Change the meaning of --max-test-loop-count=0
Previously, --max-test-loop-count=0 had no effect.  It simply enabled
fracturing which is the default.  This commit changes its behaviour to
do the opposite, essentially disable fracturing.  The execution of the
test when --max-test-loop-count=0 is limited by time only and not the
number of loop iterations.  This change gives us a fairly straight-
forward way of disabling test fracturing on the command line.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>