FW: add per-thread and per-test effective freq
Effective per-core frequency is a datapoint we (Intel) will be tracking
when we execute tests. Effective CPU frequency is displayed in MHz.

This information will be provided only in YAML logging format, when
OpenDCDiag is run with sufficient privilege to read MSR data from devfs
(root/superuser, with msr kernel module loaded), and when the retrieved
sample data appears sane.

In cases where a thread fails, assuming verbosity is set to >1 by user,
the failing thread's effective frequency may be displayed. I say "may"
because the effective frequency calculation becomes more reliable with
time. A small delta between samples may yield no result or an unreliable
one.

At the end of a test, the arithmetic mean of all test thread frequencies
will be presented.

Necessary class is only defined and implemented for Linux. A placeholder
definition exists for Windows, etc.

Signed-off-by: Joe Konno <joe.konno@intel.com>