Meson: detect if the host or target machine is x86-64
Meson follows Autotools naming and calls "host" the environment that
will run the binaries it's producing; everyone else calls "host" the
machine you're currently on. Target is the environment that will run the
binaries that are produced when run the Meson project's results (that
is, if the project is a cross-compiler, what platform it produces
binaries for). Since OpenDCDiag is not a cross-compiler, there's no
difference between host and target.

If the host/target is not x86-64, then don't include the HSW and SKX
tests and don't force the default -march= to "haswell". The default in
Meson is now the empty string, from which meson.build will make
"haswell" only if the host/target is x86-64.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>