Add non-x86 "CPU detection"
Which is a no-op. Most other architectures (with the notable exception
of Intel's own IA-64) don't have a CPUID instruction in the first
place. On Linux, you get can information from the OS by using
getauxval() for AT_HWCAP and AT_HWCAP2 and <asm/hwcap.h>. We don't do
any of that right now.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>