fw: selftests: remove the 2s timeout
A 2s timeout is currently imposed on selftests that do not specifically
request a given timeout.  This causes issues with some of the tests
whose runtimes are proportional to the number of threads in the system
under test.  These tests, e.g., selftest_timedpass_maxthreads1, cannot
declare a timeout at compile time that will work for all machines, so
it's best just to get rid of the artificial 2s limit.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>