FW: Don't change the affinity of the control thread just to set its name
We used to do that before March 2021, but thought better of it.
Unfortunately, I accidentally brought it back when I wrote the
abstracted sysdeps/*/cpu_affinity.cpp when I moved the thread naming
functionality into pin_to_logical_processor().

As a consequence, on Unix systems, run_threads() starts each of the
worker threads with the same affinity, so they're all competing for CPU
time on the same processor. They'll eventually move themselves off to
another CPU, but not without a lot of context-switching. On systems with
100+ logical processors (such as dual-socket CLX and ICX), that's a lot
of contention.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>