Buildsystem: fix passing -march flags to the HSW and SKX libraries
Passing march_flags followed by -march=skylake-avx512 is usually
fine... except that march_flags had -mtune-skylake and that has an
effect. We don't want non-AVX512 tuning for AVX512 code, so let GCC the
best tuning for it.

For AVX2 code, we do want -mtune=skylake, as there are a lot of those
out there, both in Xeon form (Xeon D, E and W) and in client form. At
some point in the future, we may want to switch to -mtune=alderlake,
but we're not there yet.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>