ifs: only modprobe the kernel driver if it isn't already there
It's possible someone compiled a kernel with the driver built-in (not a
module) and has a system without /sbin/modprobe. Or, in a more likely
situation, it's possible OpenDCDiag is being run from inside a container
that was granted access to the already-loaded /sys interface, but the
kmod package was not installed.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>