ifs: use fork()+exec() instead of system()
system() is never a good idea: it runs the shell. And that may load
scripts from the user that are not something we'd like to depend
on. More importantly, it may print stuff, including:

    sh: /sbin/modprobe: No such file or directory

So use fork() + exec() so we can control the environment more tightly.

sys/wait.h is not present on Windows, so I moved all the #includes to
inside the #if block.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>