ifs: suppress modprobe's error if it can't find the module
It is either telling us that the module doesn't exist or that we don't
have permissions to load it. Neither is a fatal issue so we shouldn't
let this be shown in logging verbosity 1.

According to kernel developers, the module should be auto-loaded on
machines that have IFS support, so the modprobe shouldn't even be
needed. However, on the machines we tested, that autoloading was somehow
broken, so we're choosing to leave it in.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>