forkfd: implement vfork(2) support
fork() works by implementing Copy-On-Write for all pages that either the
parent or the child process write to. So if the parent process continues
running while the child is between fork(2) and execve(2), then it will
keep causing page faults and requiring the OS to duplicate those pages,
with the accompanying page table update (and a TLB flush?). This problem
is aggravated if the parent process is multithreaded, as the simple act
of running in the parent will cause those threads' stacks to cause page
faults.

On Linux, we have a workaround for this: use vfork(2), which is
implemented as a call to clone(2) with the flags CLONE_VM and
CLONE_VFORK. This has the benefit that the child process is not causing
COW on the parent. But because it is still sharing the same pages, we
can't return from the forkfd() function, so instead vforkfd() adds a
callback of the same signature as glibc's clone(2) wrapper.

Fixes https://bugreports.qt.io/browse/QTBUG-104493. Matching Qt review:
https://codereview.qt-project.org/c/qt/qtbase/+/417829

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>