fw: fixes for GA + NO_LOGGING, remove some i18n
Once upon a time there was a wish for simple i18n/l10n string lookups. Some
scaffolding went in to support this feature, but it never went anywhere. Since
it's bit-rotting, it's time to prune it.

Before pruning it away entirely, some attention is required for some
preprocessor macros, conditionally compiled blocks, and functions.

In the RESTRICTED_CMDLINE case, preprocessor macros are defined in lieu of
these functions. Before, compilation failed because the functions' symbolic
names (and only their names) were expanded, resulting in errant preprocessed
code. So, the GA and RESTRICTED_CMDLINE define combination would never
have compiled in the first place.

The l10n MSGID scaffolding is pretty easy to remove with some minor touches to
a few functions and macros, so we'll do so here.

So, the following invocation:

  meson setup -Dflavor=ga -Dcmdline=restricted

Should actually compile now. This paves the way for a future change to finish
removing the l10n/i18n feature.

Signed-off-by: Joe Konno <joe.konno@intel.com>