FW/logging: include the version of the libc we're using
In the past, we've seen some tests' effectiveness depend on the libc
version, in particular just what vector instructions memcpy(), memset(),
and memcmp() use. So it might be useful to know this when trying to
reproduce an issue later on.

Dynamic build on OpenSUSE Tumbleweed:

$ ./opendcdiag -o - --selftests -e selftest_pass | yq .os
"Linux 5.18.6-1-default, glibc 2.35"

Static build on Fedora 33, run on the same Tumbleweed:

$ ./opendcdiag -o - --selftests -e selftest_pass | yq .os
"Linux 5.18.6-1-default, glibc 2.32 (statically linked)"

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>