FW: disable --test-list-randomize for now
It doesn't do anything for now because the old, weighted test list it
would use is not present. It was also surprising that it would run only
those tests, not all the tests in the tool.

We may want to re-apply randomisation later. If you want to run a
randomised test set, you can do it with --test-list-file:

 ./opendcdiag --list-tests > /tmp/list
 ./opendcdiag --test-list-randomize --test-list-file /tmp/list

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>