FW: refactor the background scan's random time-waiting code
The way we calculate the time is different, but it should be the same
result.

Previously, to calculate T ± P%, we calculated T + rnd * T * P/100
inside the lambda. Now, we calculate T * P/100 outside the lambda. This
avoids a floating point division, which was unavoidable because of the
expression:

  random_factor * deviation_percentage / 100

because random_factor is not known at compile time and, in IEEE 754
floating point, division/multiplications are not associative (the order
of the operations matter).

Now, we perform the division in the integer side, which compilers can
optimise well. Moreover, it's performed on a compile-time constant, so
the compiler knows the result and emits it on its own. This is finally
helped by using std::chrono::microseconds directly.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>