FW: move the background scan's time constants to a namespace
We can do "using namespace" for namespaces and thus shorten the lines.

I didn't move the idle load constants because the compiler must load
load them from memory somewhere (they can't be immediates in code), and
it's possible we'll add a command-line parameter to regulate these
values in the future.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>