Meson: fix make-gitid.pl with non-Git builds
Use the proper Perl stream name we've just opened.

Commit @ca4a226aef25527cdd76775a6154435cf9331f40 (PR #114) made the
script chdir() to the $MESON_SOURCE_ROOT dir so we could find the .hash
file properly. But that required opening the output file before the
chdir() (so we could deal with relative paths the right way), so I had
to change the name of the Perl stream from F to HASH. But I missed one
update.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>