CI: adjust timing-based tests to cope with slow VMs
The "--test-list-file with duration" test was asking for a test to run
for 50 ms, but given slow VMs (like those in the GitHub CI), the test
may take along time more, like 253 ms. So instead of comparing the upper
limit based on the requested duration, let's compare it to the default
duration and ensure it's less than that (I'm giving it a 100% headroom
too).

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>