FW/Unix: disable DEBUGINFO_URLS before launching gdb for backtrace
It could be useful to get glibc debug info, but that's actually easy for
us to do from the post-crash backtrace information, as OpenDCDiag prints
the glibc version in the header. But the Linux distribution's debug
server will definitely not have information about OpenDCDiag or
OpenDCDiag-based tools, so it's preferable to get a fast answer than a
more detailed one.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>