FW/logging: refactor log_{platform_,}message
This merges the two functions for the most part and makes them use a
single writev() call to write the content.

This was originally done to fix #153, but that fails because the Linux
kernel does not guarantee atomicity of write() calls and I wouldn't care
to make a guess what other OSes do. If we want to have
log_platform_message() be thread-safe, we'll need to insert a mutex.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>