FW/Win32: fix running OpenDCDiag when the path was 32 characters
If OpenDCDiag was unpacked into a dir like
  c:\OpenDCDiag-10\opendcdiag.exe
it would fail to run with the message "GetModuleFileNameA: No error"
because the path was exactly 32 characters and no error had occurred
(though that use of perror() is also wrong because GetModuleFileName
does not set errno).

This was probably introduced during refactor, when the path was a C
array, for which sizeof(path) was a correct number. sizeof(std::string)
is an implementation detail, 32 bytes on 64-bit libstdc++ with the C++11
ABI.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>