simd.conf: update after Intel Instruction Set Extensions Manual release 046
There were no relevant changes in 045. For 046[1], we have:
* avx512vp2intersect is marked as "Tiger Lake (not currently supported in any other processors)"
* raoint, prefetchiti, amxfp16, cmpccxadd, and three AVX extensions announced
* the first CPUID bits in EDX for leaf 7, subleaf 1
* GNR, GRR, SRF and MTL announced

Raptor Lake wasn't on this list, but since it's already on the Ark as
"13th Generation"[2], I added it as a commented line. Ditto for Emerald
Rapids, which was announced at Investor Day 2022. I've only added them
as comments because they appear to have no new CPUID feature bits.

The ISE didn't include the name of the E-core in GRR and SRF, but the
perfmon update[3] for MTL named "Crestmont". Since GRR and SRF are
listed to have instructions that MTL doesn't, that either means the ISE
is wrong or those two are using a generation further compared to MTL's
E-cores. I'm going on a limb here and assume the instructions listed for
them are CMT instructions and that MTL will have them, because otherwise
CMT == GRT. We may know more when the SDE tool[4] updates.

[1] https://cdrdv2.intel.com/v1/dl/getContent/671368
[2] https://ark.intel.com/content/www/us/en/ark/products/codename/215599/products-formerly-raptor-lake.html
[3] https://download.01.org/perfmon/MTL/
[4] https://www.intel.com/content/www/us/en/developer/articles/tool/software-development-emulator.html

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>