simd.conf: update with the Cove core names
After the Skylake microarchitecture, the Intel P-cores have had "Cove" names:
      Cannon Lake                    Palm Cove
      Ice Lake                       Sunny Cove
      Tiger Lake                     Willow Cove
      Alder Lake, Sapphire Rapids    Golden Cove

Sapphire Rapids was announced to have AVX512FP16 and some Alder Lake
parts whose E-cores were disabled got the feature too, so it's safe to
assume it's got the same P-core. Granite Rapids' core is not announced
anywhere that I can see, so I've kept it at GLC, though that's likely
wrong and will need to be corrected in the future.

I've kept the AMX instructions in the SPR and GNR listing instead of the
Cove because, like pconfig before it (for MKTME), they're likely to be
server-only. This may not be a good idea in the long run; I'll need to
update the script to understand mandatory and optional instructions.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>