Fix the behavior of --test-delay option.
Unlike background scan mode, --test-delay is an interactive option. If a
test is skipped, the framework will constantly attempt to restart it
without any delay. Consider the following command-line:

$ opendcdiag -vv -T 24h -e ifs --test-delay 30000

Instead of attempting to run the test every 5 minutes, the above will be
constantly trying to restart the test (if it was skipped the first time
or when it will be skipped later).

Signed-off-by: Arzhan Kinzhalin <arzhan.i.kinzhalin@intel.com>