fw: Add test_failure_package_only flag
This new flag indicates to the framework that a test failure cannot
be attributed to a specific thread or core but that it can be
attributed to a package.  Tests that request this flag are still
re-run on failure but the framework analyzes these failures and
reports on them in a different way.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>