FW/Linux: set the malloc arena count to 2x the number of processors
We do want to have an arena per thread that we will start, to avoid
contention in malloc() of every thread. Instead, it's much better to
simply allow each thread to have its own arena.

glibc's ptmalloc implementation creates a certain number of arenas
("M_ARENA_TEST", which defaults to 8 on 64-bit systems) before querying
the system config. Up until glibc 2.36, the query had a bug[1] that
failed to initialise some memory, so glibc concluded that the system had
lots of threads available, which albeit incorrect, was actually good for
us: OpenDCDiag sets the affinity for the worker threads to exactly 1
logical processor.

[ChangeLog][Framework] Fixed an issue encountered with glibc version
2.35.1 or later, which caused memory allocations in test threads to
contend with each other. This caused some tests to have reduced
efficiency in detecting SDEs.

[1] https://sourceware.org/bugzilla/show_bug.cgi?id=28850

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>