FW/topology: fix inverted conditional in parsing /proc/cpuinfo
We wanted to increase the vector holding the information about each CPU
when we found a new CPU, so it should have been >= the current size. It
was working *mostly* fine before this, so long as the Linux processor
numbers didn't have any gaps, because the next number would be equal to
the vector size. But it could crash the moment a gap existed because we
wouldn't resize the vector.

[ChangeLog][Framework] Fixed a bug parsing /proc/cpuinfo on Linux
whenever the /proc file had a gap in the processor numbering. This bug
could cause the tool to crash on start, but should not cause it to
report CPU failures.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>