Fixing average cpu load calculation on Windows
Earlier we were initializing loadavg_ atomic with infinity to assume
system is idle by default when there are no performance counters we can
use for estimation. This caused loadavg_ to always be stuck at infinity.
We should only set it to infinity when initialization of performance
counters fail, to match the behavior we have on Linux.

Signed-off-by: Wlazlyn, Patryk <patryk.wlazlyn@intel.com>