Restrict range of possible values for average cpu load on Windows
In the performance counters callback function we check how much time
really passed since we were called last time to adjust for "missed"
event. This, however  may result in huge "average load" values if we
were not called back for a long time (for example because the system was
busy or it was suspended/hibernated).  This patch provides a clamp of
average load to reasonable values.

Signed-off-by: Wlazlyn, Patryk <patryk.wlazlyn@intel.com>