FW/logging: split the crash / OS error result from the result: line
The other result types were always a string, so some parsers failed to
properly extract the details here and ended up blank.

[ChangeLog][Framework] Changed how the YAML output format logs crashes
and other unexpected operating system errors. Previously, they were a
YAML object in the "results" entry, instead of the "pass" or "fail"
string. Now, those will print either "crash" or "operating system error"
for that entry, with the details moved to the "result-details" entry.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>