IFS: Fix driver error code detection
The details file reports the hex value of the SCAN_STATUS MSR.
Previous implementation assumes 64 bits will report only 0xFD
or 0xFE when error occurs. But, according to volume 4 of the
Intel Software Developer’s Manual, only bits 39:32 are
reserved for error codes.

On this fix, I use a mask to retrieve the error codes if
present and compare them to the ones we expect.

Signed-off-by: Athenas Jimenez <athenas.jimenez.gonzalez@intel.com>