IFS: Skip test execution when during IFS wait time
Every time a test fails, OpenDCdiag will do some retries as a
way to confirm the failure. This is not applicable for IFS
test, which has a wait time between executions.

To prevent doing the whole execution during wait time, IFS test
checks for error code 0x6 and skips if present.

[ChangeLog] Skip test execution when during IFS wait time.

Signed-off-by: Athenas Jimenez <athenas.jimenez.gonzalez@intel.com>