FW/signals: remove the weird set up I devised to use a single uint32_t
In favour of a uint64_t, which is atomic in any 64-bit platform anyway.
This allows us not to depend on the details of how the OS uses
termination signals.

We also don't need the signal handler to be atomic, if we mask all the
signals that that handler would be invoked for, if it is getting run.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>