CI/Bats: check that gdb works instead of just checking its presence
In some container environments, gdb doesn't work at all. In our
self-hosted runners, I'm guessing it fails to ptrace the crashing
process, because there's no output from it in the log. But even when it
doesn't work, on Unix and x86-64, we do have the built-in register dump,
so we should be able to test that.

Meanwhile, in the hosted GitHub Actions runners, gdb hangs forever and
the signal context doesn't seem to get transmitted at all to the parent
process. In this case, not even the built-in register dump works.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>