CI: Partially revert "CI/Bats: fix the backtrace selftest again"
This partially reverts commit efbb4e0c454125f3030e618190e3e0bd154abc62,
restoring the requirement to test gdb if gdb is found to be working, but
changing what we test of gdb. Instead of trying to start and run a
process, we instead ask gdb to attach to a running process, which is
what our child_debug.cpp does.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>