FW/logging: allow for memcmp_offset() failing to find a difference
memcmp_or_fail() was introduced to allow easy checking that a given data
set was correct, reporting it in a standard format in case it isn't
(that includes virtual and physical addresses, in case they are
important for debugging). It includes an memcmp(), so we don't call into
the framework unless a difference was found. Because memcmp_or_fail() is
inline, the compiler often inlines memcmp() itself, replacing what would
be a function call into libc with a simpler version.

Interestingly, the mere act of memcmp()ing two data blocks sometimes
triggers a Silent Data Error. Therefore, we can't be assured that
memcmp() and memcmp_offset() will agree. This commit removes the
assertion that memcmp_offset() is able to find a difference and
introduces proper reporting.

    messages:
    - level: error
      data-miscompare:
       description: ''
       type:        uint8_t
       offset:      null
       address:     '0x7efd96001e00'
       actual:      null
       expected:    null
       mask:        null
       remark:      'memcmp_offset() could not locate difference'

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>