FW/random: move the thread's seeding into the thread itself
Instead of doing it serially before the child is even started. That
removes them from the critical path. It wasn't a lot of time, but it's a
nice benefit, especially when you have 960 logical processors (8 sockets
of 60 hyperthreaded cores).

This additionally makes it so each thread has a generator state that
doesn't depend on how many threads started before it. Previously, if you
decided to skip some logical processors in re-testing, the ones you did
test wouldn't get the same sequence.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>