CI: enable Clang builds
Intel doesn't use it for our internal software, because the nature of
the tests in question means they're very sensitive to code
generation. But others may want to use Clang, so this enables for them.

This is built using libstdc++ on Linux. libc++ can build it, but the
tests using pthread_cancel() seem to crash, something I can't reproduce
on FreeBSD, so I'm just chalking this up to an improper build of libc++
or libc++abi on Debian.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>