FW/logging: rip out the --use-predictable-file-names option
This was added during the port to Windows, when we thought that we'd
need to have predictable file names so the child process would be able
to reopen the files. But it turns out that Windows and the C runtime can
let child processes inherit file descriptors, so this feature was debug-
only.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>