Triage function improved for handling multiple packages with slicing.
If no CPUs are enabled for particular package, the package is not being
triaged (if the fault is isolated to this particular package).
As a result, the test doesn't success (with "no test executed" skip reason)
and doesn't introduce any confusion and unexpected reports suggesting
regressions.
Current pattern of triaged packages is: 2+..+n, ..., n, 1. For 2S (or
any setups with 2 packages) the pattern is reduced to 2, 1.
Proper number of test runs is logged (1/2/3/.. instead of 2/4/6).

Signed-off-by: Jarek Poswiata <jaroslaw.poswiata@intel.com>