FW/child_debug: use open_memstream() for the register dump FILE* stream
We can do buffering on the parent process, so this avoids the huge
sequence of write() system calls that you'd see on strace(1). Now you
just see

```
[pid 275104] write(7</memfd:>(deleted), "Registers:\n rax   = 0x0000000000"..., 4472) = 4472
```

This is done in preparation of removing the FILE* we kept per thread.

open_memstream() was a glibc extension to POSIX, but was added to the
2008 edition, so it's available everywhere (except Windows, but this
code isn't compiled there).

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>