FW/child_debug: use futexes to make the child wait for gdb
I decided to use the thread_state variable in the main thread for
this. It's currently unused: instead of setting it to the result of the
init() function, we return that state in the process's exit status. But
even if it were in use, this state variable can be used, because either
the crash happened *in* the init() or cleanup() function or init()
returned success.

I needed to enlarge this variable to 32 bits to be able to use it in a
futex.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>