FW/child_debug: don't decrease the Unix socket buffer sizes
I'm not entirely sure how they account for the two sizes, since Unix
socket datagrams are delivered immediately to the peer, so there can be
only one buffer size total. From what I can read from the Linux kernel
sources (net/unix/af_unix.c), there's a limit on the number of datagrams
pending and the sender's SNDBUF-32 specifies the maximum datagram
length. The number of pending datagrams appears to be set from the
net.unix.max_dgram_qlen sysctl at socket creation and can't be modified.

In any case, fail the whole procedure if we can't raise the limit from
unsuitable defaults. The size of the XSAVE state including the AMX is
about 10 kB, so it isn't likely to be denied.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>