FW/logging: move the file descriptor into the per_thread_data
It didn't use to be there because we kept it together with the FILE*
stream from <stdio.h>, but since we don't use that any more, we can
move. This also removes the need to keep it in the ExecState that is
passed to the child process in execve() mode.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>