FW/SharedMemory: move SApp::ExecState's contents to SharedMemory
This is a massive search-and-replace change with no actual behaviour
difference for most of the application. The only actual changes are in
the spawn_child() and exec_mode_run() functions, which must now
communicate to each other the file descriptor to the shared memory
segment instead of that temporary state file.

This makes the shared memory segment both input and output to the child
process.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>