FW/SharedMemory: don't pass sApp->current_test_duration to the child
It's not an input variable to the test. The child process alone
calculates the test end time from other variables that are
input.

This used to be used in the sliding-window slicing code, which was
removed in 8f59e10b04ad82724d3bf2975d827f0aeb1ca022 (PR #271), to
calculate the new slice's end time. That also removed the writing to
sApp->current_test_endtime.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>