CI/Bats: add a test to confirm that we are using the cpuset threads
To confirm that we got scheduled where we expected to get scheduled, we
ask the OS where we are.

This doesn't work on Windows systems with more than 64 logical
processors due how it divides the processors into Processor Groups[1],
as the number we print from the tool does not match anything meaningful
in the OS above 64.

[2] https://learn.microsoft.com/en-us/windows/win32/procthread/processor-groups

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>