FW: remove the test sliding-slice feature (1/2: child side)
The slicing feature was added back on 2019-10-23, with the
documentation:

  "Maximum number of threads to assign to the given test. This might be
  useful when the test does not scale well for a very big number of
  cores. 0 (default) means use all available. Any other value will make
  the test_run() function to be called only for a subset of the overall
  available threads, not all of thread_count. The framework will try to
  spread out package/core usage as much as it can, when this mode is
  activated"

At that time, that was really just a limit of the number of threads: all
logical processors above that limit simply weren't tested. The sliding
window functionality was added on 2019-11-05.

The motivation for the feature is sound and still relevant: there are
tests that don't scale well, for example tests with high mutex or
spinlock contention, or tests that saturate the bandwidth between CPUs
and main memory or between multiple CPUs. However, the solution of
simply sliding a window over all the CPUs in the system is not the best.
Not only does it multiply the test run time (which can be slow!) by the
number of slices, it still keeps logical processors belonging to
different sockets in the same test.

Therefore, slicing into sequential sliding  windows is being removed. In
its place, we'll run the slices in parallel and the slice size will be
dynamically and heuristically determined, aiming at avoiding unnecessary
cross-socket talk.

This commit removes the code that implements the sliding feature in the
child process side. I've avoided unnecessary churn in this commit, so
run_slices() function now is technically incorrectly named. The next
commit will clean up in the parent side.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>