FW: merge the run_thread_slices() and run_child() functions
The child no longer performs the sliding window slicing of the system,
so I merged them. To match the new convention, I've also renamed
child_run(), which now also matches exec_mode_run(). Logically, the
combined function ought to be located above wait_for_child() in the
source, but I'm not going to break git blame for this.

I want to move the CPU feature check back into the parent process to
speed up the execution of such skipping tests, but that's a task for
another day. It used to be in the parent but I moved it back on
2021-03-19 because of logging:

    FW: move the per-test check on CPU features and timing into the child

    This means we'll launch a child for tests that will be skipped, thus
    adding a bit of overhead. But it simplifies logging (especially for
    Windows support).

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>