Meson/Linux: parse the malloc.cpp source instead of malloc.cpp.o
I want to avoid having to build the .o file before extracting malloc.o
from libc.a. To parse the source file, we need a clear marker, which I
also used to avoid having to have duplicate declarations.

Because of Meson is incapable of using backslashes in custom_target()
commands (see [1]), I had to write a shell script to execute sed, so I
decided to move the ar x command there too.

[1] https://github.com/mesonbuild/meson/issues/1564

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>